+++ datafusion/core/src/optimizer/projection_push_down.rs
@@ -32,6 +32,7 @@ use crate::optimizer::utils;
 use crate::sql::utils::find_sort_exprs;
 use arrow::datatypes::{Field, Schema};
 use arrow::error::Result as ArrowResult;
+use datafusion_expr::utils::{expr_to_columns, exprlist_to_columns};
 use datafusion_expr::Expr;
 use std::{
     collections::{BTreeSet, HashSet},
@@ -159,7 +160,7 @@ fn optimize_plan(
                         new_fields.push(field.clone());
 
                         // gather the new set of required columns
-                        utils::expr_to_columns(&expr[i], &mut new_required_columns)
+                        expr_to_columns(&expr[i], &mut new_required_columns)
                     } else {
                         Ok(())
                     }
@@ -263,7 +264,7 @@ fn optimize_plan(
                         new_window_expr.push(expr.clone());
                         new_required_columns.insert(column);
                         // add to the new set of required columns
-                        utils::expr_to_columns(expr, &mut new_required_columns)
+                        expr_to_columns(expr, &mut new_required_columns)
                     } else {
                         Ok(())
                     }
@@ -271,7 +272,7 @@ fn optimize_plan(
             }
 
             // for all the retained window expr, find their sort expressions if any, and retain these
-            utils::exprlist_to_columns(
+            exprlist_to_columns(
                 &find_sort_exprs(&new_window_expr),
                 &mut new_required_columns,
             )?;
@@ -296,7 +297,7 @@ fn optimize_plan(
             // * remove any aggregate expression that is not required
             // * construct the new set of required columns
 
-            utils::exprlist_to_columns(group_expr, &mut new_required_columns)?;
+            exprlist_to_columns(group_expr, &mut new_required_columns)?;
 
             // Gather all columns needed for expressions in this Aggregate
             let mut new_aggr_expr = Vec::new();
@@ -309,7 +310,7 @@ fn optimize_plan(
                     new_required_columns.insert(column);
 
                     // add to the new set of required columns
-                    utils::expr_to_columns(expr, &mut new_required_columns)
+                    expr_to_columns(expr, &mut new_required_columns)
                 } else {
                     Ok(())
                 }
@@ -484,7 +485,7 @@ fn optimize_plan(
         | LogicalPlan::Extension { .. } => {
             let expr = plan.expressions();
             // collect all required columns by this plan
-            utils::exprlist_to_columns(&expr, &mut new_required_columns)?;
+            exprlist_to_columns(&expr, &mut new_required_columns)?;
 
             // apply the optimization to all inputs of the plan
             let inputs = plan.inputs();
